# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DashboardRead": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "current_month_expenses_minor": {
            "title": "Current Month Expenses Minor",
            "type": "integer"
          },
          "current_month_income_minor": {
            "title": "Current Month Income Minor",
            "type": "integer"
          },
          "excluded_currency_account_count": {
            "title": "Excluded Currency Account Count",
            "type": "integer"
          },
          "goal_saved_minor": {
            "title": "Goal Saved Minor",
            "type": "integer"
          },
          "goal_target_minor": {
            "title": "Goal Target Minor",
            "type": "integer"
          },
          "net_cash_flow_minor": {
            "title": "Net Cash Flow Minor",
            "type": "integer"
          },
          "safe_to_spend_minor": {
            "title": "Safe To Spend Minor",
            "type": "integer"
          },
          "safe_to_spend_shortfall_minor": {
            "title": "Safe To Spend Shortfall Minor",
            "type": "integer"
          },
          "savings_rate_basis_points": {
            "title": "Savings Rate Basis Points",
            "type": "integer"
          },
          "total_available_balance_minor": {
            "title": "Total Available Balance Minor",
            "type": "integer"
          },
          "upcoming_bill_count": {
            "title": "Upcoming Bill Count",
            "type": "integer"
          },
          "upcoming_bills_minor": {
            "title": "Upcoming Bills Minor",
            "type": "integer"
          }
        },
        "required": [
          "currency",
          "total_available_balance_minor",
          "current_month_income_minor",
          "current_month_expenses_minor",
          "net_cash_flow_minor",
          "savings_rate_basis_points",
          "upcoming_bills_minor",
          "upcoming_bill_count",
          "goal_saved_minor",
          "goal_target_minor",
          "safe_to_spend_minor",
          "safe_to_spend_shortfall_minor",
          "excluded_currency_account_count"
        ],
        "title": "DashboardRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SafeToSpendRead": {
        "properties": {
          "available_balance_minor": {
            "title": "Available Balance Minor",
            "type": "integer"
          },
          "confirmed_income_minor": {
            "title": "Confirmed Income Minor",
            "type": "integer"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "daily_allowance_minor": {
            "title": "Daily Allowance Minor",
            "type": "integer"
          },
          "deductions": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Deductions",
            "type": "object"
          },
          "explanation": {
            "items": {
              "type": "string"
            },
            "title": "Explanation",
            "type": "array"
          },
          "minimum_untouched_minor": {
            "title": "Minimum Untouched Minor",
            "type": "integer"
          },
          "raw_safe_to_spend_minor": {
            "title": "Raw Safe To Spend Minor",
            "type": "integer"
          },
          "safe_to_spend_minor": {
            "title": "Safe To Spend Minor",
            "type": "integer"
          },
          "shortfall_minor": {
            "title": "Shortfall Minor",
            "type": "integer"
          },
          "until_date": {
            "format": "date",
            "title": "Until Date",
            "type": "string"
          }
        },
        "required": [
          "currency",
          "until_date",
          "safe_to_spend_minor",
          "raw_safe_to_spend_minor",
          "shortfall_minor",
          "minimum_untouched_minor",
          "available_balance_minor",
          "confirmed_income_minor",
          "deductions",
          "daily_allowance_minor",
          "explanation"
        ],
        "title": "SafeToSpendRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Offline-sync-ready personal finance API. All monetary values are integer minor units.",
    "title": "MoneyPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard/summary": {
      "get": {
        "operationId": "dashboard_summary_api_v1_dashboard_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Dashboard Summary",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/safe-to-spend": {
      "get": {
        "operationId": "safe_to_spend_api_v1_safe_to_spend_get",
        "parameters": [
          {
            "in": "query",
            "name": "until_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until Date"
            }
          },
          {
            "in": "query",
            "name": "confirmed_income_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Confirmed Income Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "debt_minimum_payments_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Debt Minimum Payments Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "planned_savings_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Planned Savings Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "goal_contributions_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Goal Contributions Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "emergency_reserve_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Emergency Reserve Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "credit_card_obligations_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Credit Card Obligations Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "safety_buffer_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Safety Buffer Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "known_one_time_expenses_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Known One Time Expenses Minor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "forecast_uncertainty_buffer_minor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Forecast Uncertainty Buffer Minor",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeToSpendRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Safe To Spend",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}